Server/requirements.txt
alabaster==0.7.11
amqp==2.3.2
Babel==2.6.0
bcrypt==3.1.4
billiard==3.5.0.3
celery==4.1.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
Django==2.0.5
django-cors-headers==2.4.0
docutils==0.14
gunicorn==19.9.0
idna==2.7
imagesize==1.0.0
Jinja2==2.10
kombu==4.2.1
MarkupSafe==1.0
mongoengine==0.15.0
numpy==1.14.5
packaging==17.1
pandas==0.23.1
pycparser==2.18
Pygments==2.2.0
pymongo==3.6.1
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
requests==2.19.1
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.6
sphinxcontrib-websupport==1.1.0
tornado==5.1
urllib3==1.23
vine==1.1.4
